requirements.txt
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
Pillow==10.2.0
numpy==1.26.4
pygame==2.5.2 